fix: rename commands
