+++ pkg/cmd/profile/config.go
@@ -1,20 +1,20 @@
-package config
+package profile
 
 import (
 	"github.com/MakeNowJust/heredoc"
-	"github.com/redhat-developer/app-services-cli/pkg/cmd/config/generate"
-	"github.com/redhat-developer/app-services-cli/pkg/cmd/config/manage"
-	"github.com/redhat-developer/app-services-cli/pkg/cmd/config/status"
-	"github.com/redhat-developer/app-services-cli/pkg/cmd/config/view"
 	"github.com/redhat-developer/app-services-cli/pkg/cmd/factory"
+	"github.com/redhat-developer/app-services-cli/pkg/cmd/profile/generate"
+	"github.com/redhat-developer/app-services-cli/pkg/cmd/profile/manage"
+	"github.com/redhat-developer/app-services-cli/pkg/cmd/profile/status"
+	"github.com/redhat-developer/app-services-cli/pkg/cmd/profile/view"
 
 	"github.com/spf13/cobra"
 )
 
 func NewConfigCommand(f *factory.Factory) *cobra.Command {
 	cmd := &cobra.Command{
-		Use:   "config",
-		Short: "Generates and manages services configurations",
+		Use:   "profile",
+		Short: "Generates and manages services profiles",
 		Long: heredoc.Doc(`
 		Group of command that provide abstraction for managed services configuration.
