+++ pkg/cmd/profile/generate/generate.go
