+++ pkg/cmd/profile/status/status.go
