+++ pkg/cmd/profile/view/view.go
